+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.7",
-  "contentSha256_12": "8c7ca44de83c",
+  "pkgVersion": "1.1.0-alpha.8",
+  "contentSha256_12": "76f510da535d",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.1.0-alpha.7";
-export const contentSha256_12 = "8c7ca44de83c";
+export const pkgVersion = "1.1.0-alpha.8";
+export const contentSha256_12 = "76f510da535d";
 export const file = "nodalia-cards.bundle.js";